Figured it out I think (encoding issue)

+++ app/models/dj_metrics.rb
@@ -24,18 +24,26 @@ class DjMetrics
     end
   end
 
-  def metrics_ready?
-    File.exist?(METRICS_DIR + '/ready')
-  end
+  # def metrics_ready?
+  #   File.exist?(METRICS_DIR + '/ready')
+  # end
+
+  # def clear!
+  #   Dir['/app/prometheus-metrics/*'].each do |file_path|
+  #     File.unlink(file_path)
+  #   end
+  # end
 
   def register_metrics_for_delayed_job_worker!
     Dir['/app/prometheus-metrics/*'].each do |file_path|
+      next unless file_path.match?(/_#{Process.pid}.bin/) # only delete our own pid files.
+
       File.unlink(file_path)
     end
 
     register_metrics_for_metrics_endpoint!
     refresh_queue_sizes!
-    FileUtils.touch('/app/prometheus-metrics/ready')
+    # FileUtils.touch('/app/prometheus-metrics/ready')
   end
 
   def dj_job_status_total_metric
@@ -70,10 +78,10 @@ class DjMetrics
     end
 
     # These are the ones that are now empty (if any)
-    # others.each do |queue|
-    #   Rails.logger.info "Setting #{queue} to size 0"
-    #   dj_queue_size_metric.set(0, labels: { queue: queue.encode('ascii-8bit') })
-    # end
+    others.each do |queue|
+      Rails.logger.info "Setting #{queue} to size 0"
+      dj_queue_size_metric.set(0, labels: { queue: queue.encode('ascii-8bit') })
+    end
   end
 
   private
